fix(gateway): keep lifecycle thinking active-only
